Refactor some classes that used the TodoBackboneAccess object as this one is now merged with the SilverpeasCalendar service. The deletion of todos (and hence todos' attendees) is now taken automatically into account by the Calendar service
